% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{arma_coefficients}
\alias{arma_coefficients}
\title{ARMA mean equation coefficients}
\usage{
arma_coefficients(object)
}
\arguments{
\item{object}{an object of class \dQuote{tsgarch.spec} or
\dQuote{tsgarch.estimate}.}
}
\value{
a list with elements \sQuote{ar} and \sQuote{ma}, each a numeric
vector (possibly of length zero if that component is not present).
}
\description{
Returns the AR and MA coefficients implied by the raw
Durbin-Levinson (pacf-space) parameters used internally during estimation
(see \code{\link{garch_modelspec}}). Stationarity of the AR polynomial and
invertibility of the MA polynomial are guaranteed by construction and do
not need to be checked.
}
