% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{benchmark_fcp}
\alias{benchmark_fcp}
\title{FCP GARCH Benchmark}
\usage{
benchmark_fcp(control = nloptr_fast_options())
}
\arguments{
\item{control}{control arguments for the nloptr solver.}
}
\value{
An object of class \dQuote{benchmark.fcp} which has a \dQuote{as_flextable}
method for nice printing of the results.
}
\description{
The GARCH(1,1) FCP benchmark.
}
\details{
The benchmark of Fiorentini et al. (1996) on the Deutsche Mark British Pound
returns is based on a GARCH(1,1) model with a constant in the conditional mean
equation, and normally distributed errors.
}
\references{
\insertRef{Fiorentini1996}{tsgarch}
}
