% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{benchmark_laurent}
\alias{benchmark_laurent}
\title{Laurent APARCH Benchmark}
\usage{
benchmark_laurent(control = nloptr_fast_options())
}
\arguments{
\item{control}{control arguments for the nloptr solver.}
}
\value{
An object of class \dQuote{benchmark.aparch} which has a \dQuote{as_flextable}
method for nice printing of the results.
}
\description{
The APARCH(1,1) benchmark of Laurent (2003).
}
\details{
The benchmark of Laurent (2003) on the Nikkei daily log returns is
based on an APARCH(1,1) model with a constant in the conditional mean
equation, and normally distributed errors.
}
\references{
\insertRef{Laurent2004}{tsgarch}
}
