% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{BIC.tsgarch.estimate}
\alias{BIC.tsgarch.estimate}
\alias{BIC}
\title{Bayesian Information Criterion}
\usage{
\method{BIC}{tsgarch.estimate}(object, ...)
}
\arguments{
\item{object}{an object of class \dQuote{tsgarch.estimate}.}

\item{...}{not currently used.}
}
\value{
A numeric value.
}
\description{
Extract the BIC from an estimated model.
}
