% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandwich.R
\name{bread.tsgarch.estimate}
\alias{bread.tsgarch.estimate}
\title{Bread Method}
\usage{
\method{bread}{tsgarch.estimate}(x, ...)
}
\arguments{
\item{x}{an object of class \dQuote{tsgarch.estimate}.}

\item{...}{not currently used.}
}
\value{
The analytic hessian of the model.
}
\description{
Bread Method
}
\author{
Alexios Galanos
}
