% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dmbp}
\alias{dmbp}
\title{Deutschemark/British pound Exchange Rate}
\format{
\subsection{\code{dmbp}}{

A data.frame containing 2x1974 observations
\describe{
\item{rate}{The exchange rate}
\item{monday}{Dummy indicator (see descriptiom)}
}
}
}
\source{
Journal of Business & Economic Statistics Data Archive
}
\usage{
dmbp
}
\description{
The Bollerslev-Ghysel benchmark dataset. The variables in the data set are
the daily percentage nominal returns computed as 100 \deqn{ln(Pt) - ln(Pt-1)},
where Pt is the bilateral Deutschemark/British pound rate constructed from
the corresponding U.S. dollar rates, and a dummy variable that takes the
value of 1 on Mondays and other days following no trading in the Deutschemark
or British pound/ U.S. dollar market during regular European trading hours,
and 0 otherwise.
The data spans the period from 1984-01-03 through 1991-12-31, but exact dates
are not known as this dataset did not provide an index.
This dataset is included as it is used for the GARCH benchmark.
}
\keyword{datasets}
