% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandwich.R
\name{estfun.tsgarch.estimate}
\alias{estfun.tsgarch.estimate}
\title{Score Method}
\usage{
\method{estfun}{tsgarch.estimate}(x, ...)
}
\arguments{
\item{x}{an object of class \dQuote{tsgarch.estimate}.}

\item{...}{not currently used.}
}
\value{
The score matrix
}
\description{
Score Method
}
\details{
The function returns the scores of the negative of the log likelihood
at the optimal solution. The scores are from the second pass of the optimizer
using scaling and hence represent the scaled scores. These are then rescaled
before returning the matrix.
}
\author{
Alexios Galanos
}
