% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{fitted.tsgarch.estimate}
\alias{fitted.tsgarch.estimate}
\alias{fitted}
\alias{fitted.tsgarch.multi_estimate}
\title{Extract Model Fitted Values}
\usage{
\method{fitted}{tsgarch.estimate}(object, ...)

\method{fitted}{tsgarch.multi_estimate}(object, ...)
}
\arguments{
\item{object}{an object of class \dQuote{tsgarch.estimate} or
\dQuote{tsgarch.multi_estimate}.}

\item{...}{not currently used.}
}
\value{
An xts vector of the fitted values (the model's conditional mean)
for the univariate type objects, always the same length as the original
series. When \sQuote{arma = c(0,0)} (the default) this is a constant,
equal to \sQuote{mu} (or zero if \sQuote{constant = FALSE}). When an ARMA
mean equation is specified (currently only for \sQuote{model = "garch"},
\sQuote{"egarch"}, \sQuote{"gjrgarch"}, \sQuote{"aparch"}, \sQuote{"fgarch"}
or \sQuote{"cgarch"}) this is the genuinely time-varying conditional mean
implied by the AR/MA recursion; see \code{\link{arma_coefficients}} for
the estimated AR/MA coefficients. In the case of a multi-estimate object,
a list of xts vectors is returned if the individual univariate objects
have unequal indices, else an xts matrix is returned.
}
\description{
Extract the fitted values of the estimated model.
}
