% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specification.R
\name{garch_modelspec}
\alias{garch_modelspec}
\title{GARCH Model Specification}
\usage{
garch_modelspec(
  y,
  model = "garch",
  constant = FALSE,
  order = c(1, 1),
  arma = c(0, 0),
  variance_targeting = FALSE,
  vreg = NULL,
  multiplicative = FALSE,
  init = c("unconditional", "sample", "backcast"),
  backcast_lambda = 0.7,
  sample_n = 10,
  distribution = "norm",
  ...
)
}
\arguments{
\item{y}{an xts vector with either \dQuote{Date} or \dQuote{POSIXct} index.}

\item{model}{the type of GARCH model. Valid choices are \dQuote{garch} for
vanilla GARCH, \dQuote{gjrgarch} for asymmetric GARCH, \dQuote{egarch} for
exponential GARCH, \dQuote{aparch} for asymmetric power ARCH,
\dQuote{fgarch} for Family GARCH, \dQuote{cgarch} for component GARCH,
\dQuote{igarch} for integrated GARCH, and \dQuote{ewma} for the EWMA model.}

\item{constant}{whether to estimate a constant (mean) for y,}

\item{order}{the (p,q) GARCH order.}

\item{arma}{the (ar,ma) order of the ARMA mean equation, jointly estimated
with the GARCH variance equation. Defaults to \sQuote{c(0,0)} (no ARMA
dynamics in the mean, matching prior behavior where the mean is either
zero or a constant). Supported for all models except \dQuote{igarch} and
\dQuote{ewma} (which share the plain \dQuote{garch} mean equation with no
native ARMA support). Stationarity of the AR polynomial and invertibility
of the MA polynomial are guaranteed by construction via a Durbin-Levinson
(partial autocorrelation) reparameterization, and therefore do not
require additional nonlinear constraints during estimation. The estimated
AR/MA coefficients can be extracted with \code{\link{arma_coefficients}};
see also \code{\link{fitted}} for the (possibly time-varying) conditional
mean and \code{\link{residuals}} for the ARMA innovations.}

\item{variance_targeting}{whether to use variance targeting rather than
estimating the conditional variance intercept.}

\item{vreg}{an optional xts matrix of regressors in the conditional variance
equation.}

\item{multiplicative}{whether to exponentiate the contribution of the
regressors else will be additive. In the case of the \dQuote{egarch} model,
since this is already a multiplicative model, the regressors are additive
irrespective of the choice made.}

\item{init}{the method to use to initialize the recursion of the conditional
variance.}

\item{backcast_lambda}{the decay power for the exponential smoothing used
when initializing the recursion using the backcast method.}

\item{sample_n}{the number of data points to use when initializing the
recursion using the sample method.}

\item{distribution}{a valid distribution from the available
re-parameterized distributions of the package.}

\item{...}{not used.}
}
\value{
An object of class \dQuote{tsgarch.spec}.
}
\description{
Specifies a GARCH model prior to estimation.
}
\details{
The specification object holds the information and data which is
then passed to the maximum likelihood estimation routines.
}
\author{
Alexios Galanos
}
