% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidymodel.R
\name{glance.tsgarch.estimate}
\alias{glance.tsgarch.estimate}
\alias{glance}
\title{Construct a summary glance of an estimated GARCH object}
\usage{
\method{glance}{tsgarch.estimate}(x, ...)
}
\arguments{
\item{x}{an object of class \dQuote{tsgarch.estimate}.}

\item{...}{not currently used.}
}
\value{
A tibble().
}
\description{
glance method for tsgarch
}
