% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{halflife.tsgarch.estimate}
\alias{halflife.tsgarch.estimate}
\alias{halflife}
\title{Half Life}
\usage{
\method{halflife}{tsgarch.estimate}(object, ...)
}
\arguments{
\item{object}{an object.}

\item{...}{not currently used.}
}
\value{
a numeric value representing the half life in periods based on the
frequency of the underlying data.
}
\description{
Calculates and returns the half-life of a model.
}
\details{
The half life is defined as the period it
takes a series  to reach half its long-term average values. For a GARCH model
this is defined as \eqn{log(0.5)/log(P)} where P is the persistence.
}
