% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{newsimpact}
\alias{newsimpact}
\alias{newsimpact.tsgarch.estimate}
\title{News Impact Curve}
\usage{
newsimpact(object, epsilon = NULL, ...)

\method{newsimpact}{tsgarch.estimate}(object, epsilon = NULL, ...)
}
\arguments{
\item{object}{an object.}

\item{epsilon}{a user supplied zero mean noise vector. If this is NULL
then a vector is created from the actual data using the minimum and maximum
range.}

\item{...}{additional parameters passed to the method.}
}
\value{
An object of class \dQuote{tsgarch.newsimpact}.
}
\description{
General method the news impact of a model
}
