% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{omega}
\alias{omega}
\alias{omega.tsgarch.estimate}
\alias{omega.tsgarch.spec}
\title{Omega (Variance Equation Intercept)}
\usage{
omega(object, ...)

\method{omega}{tsgarch.estimate}(object, ...)

\method{omega}{tsgarch.spec}(object, ...)
}
\arguments{
\item{object}{an object.}

\item{...}{additional parameters passed to the method.}
}
\value{
a numeric value representing the value of the intercept.
}
\description{
Returns the intercept of a GARCH model.
}
\details{
The intercept is either estimated directly as part of the model else
indirectly if variance targeting was selected.
}
