% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{persistence}
\alias{persistence}
\alias{persistence.tsgarch.estimate}
\alias{persistence.tsgarch.spec}
\title{Model Persistence}
\usage{
persistence(object, ...)

\method{persistence}{tsgarch.estimate}(object, ...)

\method{persistence}{tsgarch.spec}(object, ...)
}
\arguments{
\item{object}{an object of some class.}

\item{...}{additional parameters passed to the method.}
}
\value{
The persistence of the estimated model. For GARCH models, the formulation
varies by the type of model. See the vignette for more details.
}
\description{
General method the persistence of a model.
}
