% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{pit.tsgarch.estimate}
\alias{pit.tsgarch.estimate}
\alias{pit}
\title{Probability Integral Transform (PIT)}
\usage{
\method{pit}{tsgarch.estimate}(object, ...)
}
\arguments{
\item{object}{an object.}

\item{...}{not currently used.}
}
\value{
An xts vector of the conditional probabilities.
}
\description{
Calculates and returns the conditional probability integral
transform given the data and estimated density
}
\details{
The PIT is essentially the probabilities returned from the cumulative
distribution function (*p) given the data and estimated value of the mean,
conditional standard deviation and any other distributional parameters.
}
