% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.tsgarch.newsimpact}
\alias{plot.tsgarch.newsimpact}
\title{News Impact Plot}
\usage{
\method{plot}{tsgarch.newsimpact}(x, y = NULL, ...)
}
\arguments{
\item{x}{an object of class \dQuote{tsgarch.newsimpact}.}

\item{y}{not used.}

\item{...}{not used.}
}
\description{
Plot method for newsimpact class.
}
