% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{as_flextable}
\alias{bread}
\alias{estfun}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{flextable}{\code{\link[flextable:as_flextable]{as_flextable()}}}

  \item{sandwich}{\code{\link[sandwich:bread]{bread()}}, \code{\link[sandwich:estfun]{estfun()}}}
}}

