% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{sigma.tsgarch.estimate}
\alias{sigma.tsgarch.estimate}
\alias{sigma}
\title{Extract Volatility (Conditional Standard Deviation)}
\usage{
\method{sigma}{tsgarch.estimate}(object, ...)
}
\arguments{
\item{object}{an object of class \dQuote{tsgarch.estimate}, \dQuote{tsgarch.predict}
or \dQuote{tsgarch.simulate}.}

\item{...}{not currently used.}
}
\value{
An xts vector of the conditional volatility.
}
\description{
Extract the conditional standard deviation from a GARCH model.
}
