% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.tsgarch.estimate}
\alias{summary.tsgarch.estimate}
\alias{summary}
\title{GARCH Model Estimation Summary}
\usage{
\method{summary}{tsgarch.estimate}(object, digits = 4, vcov_type = "H", include_persistence = TRUE, ...)
}
\arguments{
\item{object}{an object of class \dQuote{tsgarch.estimate}.}

\item{digits}{integer, used for number formatting. Optionally, to avoid
scientific notation, set \sQuote{options(scipen=999)}.}

\item{vcov_type}{the type of standard errors based on the vcov estimate (see \code{\link{vcov}}).}

\item{include_persistence}{whether to include the estimate of the persistence and
its calculated standard errors (calculated using the \code{\link[TMB]{sdreport}})
in the output.}

\item{...}{not currently used.}
}
\value{
A list with summary information of class \dQuote{summary.tsgarch.estimate}.
}
\description{
Summary method for class \dQuote{tsgarch.estimate}
}
