% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{summary.tsgarch.profile}
\alias{summary.tsgarch.profile}
\title{GARCH Profile Summary}
\usage{
\method{summary}{tsgarch.profile}(object, digits = 4, measure = "RMSE", ...)
}
\arguments{
\item{object}{an object of class \dQuote{tsgarch.profile}.}

\item{digits}{integer, used for number formatting. Optionally, to avoid
scientific notation, set \sQuote{options(scipen=999)}.}

\item{measure}{either one of the included measure in the summary slot of
the returned object, which currently includes the relative error measures
\dQuote{RMSE}, \dQuote{MAE}, \dQuote{MAPE}, summary measures on the estimated
values \dQuote{MEAN}, \dQuote{MEDIAN}, \dQuote{P20} and \dQuote{P80},
else any other user calculated measure which has been generated in the summary table
post processing.}

\item{...}{not currently used.}
}
\value{
A list with summary information of class \dQuote{summary.tsgarch.profile}.
}
\description{
Summary method for class \dQuote{tsgarch.profile}
}
