% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidymodel.R
\name{tidy.tsgarch.estimate}
\alias{tidy.tsgarch.estimate}
\alias{tidy}
\title{Tidy an estimate GARCH object}
\usage{
\method{tidy}{tsgarch.estimate}(x, conf.int = FALSE, conf.level = 0.95, vcov_type = "H", ...)
}
\arguments{
\item{x}{an object of class \dQuote{tsgarch.estimate}.}

\item{conf.int}{Logical indicating whether or not to include a confidence
interval in the tidied output. Defaults to FALSE.}

\item{conf.level}{The confidence level to use for the confidence interval if
conf.int = TRUE. Must be strictly greater than 0 and less than 1. Defaults to
0.95, which corresponds to a 95 percent confidence interval.
scientific notation, set \sQuote{options(scipen=999)}.}

\item{vcov_type}{the vcov method to use.}

\item{...}{not currently used.}
}
\value{
A tibble() with columns:
\itemize{
\item{conf.high Upper bound on the confidence interval for the estimate.}
\item{conf.low Lower bound on the confidence interval for the estimate.}
\item{estimate The estimated value of the regression term.}
\item{p.value The two-sided p-value associated with the observed statistic.}
\item{statistic The value of a T-statistic to use in a hypothesis that the
regression term is non-zero.}
\item{std.error The standard error of the regression term.}
\item{term The name of the regression term.}
}
}
\description{
Tidy for tsgarch
}
