% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multispec.R
\name{to_multi_estimate}
\alias{to_multi_estimate}
\title{Convert a list of tsgarch.estimate objects to a multi_estimate object}
\usage{
to_multi_estimate(object, ...)
}
\arguments{
\item{object}{an list with \dQuote{tsgarch.estimate} objects.}

\item{...}{none}
}
\value{
a validated object of class \dQuote{tsgarch.multi_estimate}.
}
\description{
Convert a list of tsgarch.estimate objects to a multi_estimate object
}
\details{
This is a convenience method which provides the flexibility
to manually estimate each series and then convert the list of these objects
to a multi-estimation object, rather than having to use the estimate method
on a multi-specification object.
}
\author{
Alexios Galanos
}
