% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{tsequation.tsgarch.estimate}
\alias{tsequation.tsgarch.estimate}
\alias{tsequation}
\title{Model Equation (LaTeX)}
\usage{
\method{tsequation}{tsgarch.estimate}(object, ...)
}
\arguments{
\item{object}{an object of class \dQuote{tsgarch.estimate}.}

\item{...}{not currently used.}
}
\value{
A list of equations in LaTeX which can be used in documents. This is
a list with 4 slots for the conditional distribution, the conditional volatility,
the persistence and unconditional variance equations.
}
\description{
Generates a list of model equations in LaTeX.
}
\details{
This method is called in the summary when the format output option
chosen is \dQuote{flextable}.
}
