% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{arma_coefficients}
\alias{arma_coefficients}
\title{ARMA mean equation coefficients}
\usage{
arma_coefficients(object)
}
\arguments{
\item{object}{an object of class \dQuote{tsgarch.spec} or
\dQuote{tsgarch.estimate}.}
}
\value{
a list with elements \sQuote{ar} and \sQuote{ma}, each a numeric
vector (possibly of length zero if that component is not present).
}
\description{
Returns the AR and MA coefficients of the estimated (or
specified) ARMA mean equation (see \code{\link{garch_modelspec}}).
Ordinarily these are recovered from the raw Durbin-Levinson (pacf-space)
parameters used internally during estimation, with stationarity of the AR
polynomial and invertibility of the MA polynomial guaranteed by
construction. If, however, an entire AR and/or MA polynomial has been
fixed (every \sQuote{arpacf}/\sQuote{mapacf} row has \sQuote{estimate ==
0}; see \code{\link{garch_modelspec}}), the corresponding \sQuote{value}
entries are themselves the target ar/ma coefficients and are returned
as-is.
}
