% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arma.R
\name{arma_inverse_roots}
\alias{arma_inverse_roots}
\title{Inverse AR/MA Roots}
\usage{
arma_inverse_roots(object)
}
\arguments{
\item{object}{an object of class \dQuote{tsgarch.estimate} or
\dQuote{tsgarch.spec}.}
}
\value{
A list with elements \sQuote{ar} and \sQuote{ma}, each a complex
vector of inverse roots. The AR polynomial is
\eqn{1 - \phi_1 z - \dots - \phi_p z^p}{1 - phi_1 z - ... - phi_p z^p} and
the MA polynomial is
\eqn{1 + \theta_1 z + \dots + \theta_q z^q}{1 + theta_1 z + ... + theta_q z^q}.
If either component is absent, the corresponding vector is \code{complex(0)}.
}
\description{
Computes the inverse roots of the AR and MA polynomials of the
mean equation in a fitted or specified ARMA-GARCH model.
}
