% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{as_flextable.benchmark.fcp}
\alias{as_flextable.benchmark.fcp}
\title{Transform an object into flextable}
\usage{
\method{as_flextable}{benchmark.fcp}(x, digits = 4, ...)
}
\arguments{
\item{x}{an object of class \dQuote{benchmark.fcp}.}

\item{digits}{integer, used for number formatting. Optionally, to avoid
scientific notation, set \sQuote{options(scipen=999)}.}

\item{...}{additional arguments passed to flextable method.}
}
\value{
A flextable object.
}
\description{
Transforms a \dQuote{benchmark.fcp} object into a flextable.
}
