% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{fitted.tsgarch.estimate}
\alias{fitted.tsgarch.estimate}
\alias{fitted}
\title{Extract Model Fitted Values}
\usage{
\method{fitted}{tsgarch.estimate}(object, ...)
}
\arguments{
\item{object}{an object of class \dQuote{tsgarch.estimate}.}

\item{...}{not currently used.}
}
\value{
An xts vector of the fitted values. Since only a constant is supported
in the conditional mean equation this is either a vector with a constant else
a vector with zeros.
}
\description{
Extract the fitted values of the estimated model.
}
