% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.tsgarch.estimate}
\alias{plot.tsgarch.estimate}
\title{Estimated Model Plots}
\usage{
\method{plot}{tsgarch.estimate}(x, y = NULL, ...)
}
\arguments{
\item{x}{an object of class \dQuote{tsgarch.estimate}.}

\item{y}{not used.}

\item{...}{not used.}
}
\value{
a panel with plots for the estimated sigma value, the news impact curve
and a \dQuote{QQ} plot of the standardized residuals.
}
\description{
Plot method for \dQuote{tsgarch.estimate} class.
}
