% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{tsfilter.tsgarch.estimate}
\alias{tsfilter.tsgarch.estimate}
\alias{tsfilter}
\alias{tsfilter.tsgarch.spec}
\title{Model Filtering}
\usage{
\method{tsfilter}{tsgarch.estimate}(object, y = NULL, newxreg = NULL, newvreg = NULL, ...)

\method{tsfilter}{tsgarch.spec}(object, y = NULL, newxreg = NULL, newvreg = NULL, ...)
}
\arguments{
\item{object}{an object of class \dQuote{tsgarch.estimate} or \dQuote{tsgarch.spec}.}

\item{y}{an xts vector of new values to filter. Can also be NULL in which case the}

\item{newxreg}{not currently used,}

\item{newvreg}{variance regressors with the same number of rows as y. This can be either
a numeric or xts matrix. Only needed if the model was estimated with regressors in the
variance equation.}

\item{...}{additional arguments for future expansion options.}
}
\value{
A \dQuote{tsgarch.estimate} object with updated information.
}
\description{
Filters new data based on an already estimated model.
}
\details{
The method filters new data and updates the object with this new information
so that it can be called recursively as new data arrives. It is also possible to use
a specification object with fixed parameters, by appropriately setting the values
of the parmatrix object in the specification slot. In this case, the returned object
will also be of classs \dQuote{tsgarch.estimate} but with missing information
regarding the hessian and scores. However, it will still be possible to use the
returned object for prediction.
For the argument y, this should strictly be new data coming after the last
time index of the existing object data. The existing dataset and y are merged
on their time indices and only information coming after the existing data set
time index is appended. No replacement of existing data is possible using this
approach (the user can directly replace the existing data or parts of manually
for this type of use case).
}
