% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsgarch-package.R
\docType{package}
\name{tsgarch-package}
\alias{tsgarch}
\alias{tsgarch-package}
\title{tsgarch: Univariate GARCH Models}
\description{
Multiple flavors of GARCH with a rich assortment of distributions. Methods for specification, estimation, prediction, filtering, statistical testing and much more. Represents a partial re-write and re-think of rugarch, 10 years later.
}
\keyword{internal}
