% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{unconditional.tsgarch.estimate}
\alias{unconditional.tsgarch.estimate}
\alias{unconditional}
\title{Unconditional Value}
\usage{
\method{unconditional}{tsgarch.estimate}(object, ...)
}
\arguments{
\item{object}{an object of class \dQuote{tsgarch.estimate}.}

\item{...}{not currently used.}
}
\value{
A numeric vector of length 1 of the unconditional variance of the model.
}
\description{
Unconditional value of a GARCH model variance.
}
